refactor: re-organized imports and exports for components
