+++ src/components/atoms/navigation/NavBrand.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Navbar as BSNavbar } from 'react-bootstrap';
-import LogoNavbar from 'rifui/LogoNavbar';
+import { LogoNavbar } from 'rifui';
 
 const NavBrand = () => (
   <BSNavbar.Brand>
