+++ src/rifui/components/atoms/LogoNavbar.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import logo from './assets/logo-full.svg';
-import logo_full_blk from './assets/logo-full-blk.svg';
+import logo_new from '../../assets/logo_new.svg';
 
 interface ILogo extends React.Props<HTMLImageElement> {
   alt?: string;
+  height?: string | number;
   style?: React.CSSProperties;
 }
 
@@ -11,7 +11,7 @@ export default React.forwardRef(function LogoFooter(
   props: ILogo,
   ref?: React.Ref<HTMLImageElement> | null,
 ) {
-  const { alt = 'RIF OS', ...other } = props;
+  const { alt = 'RIF OS', height = '40px', ...other } = props;
 
-  return <img src={logo_full_blk} alt={alt} ref={ref} {...other} />;
+  return <img src={logo_new} alt={alt} height={height} ref={ref} {...other} />;
 });
