+++ src/rifui/components/atoms/buttons/index.ts
@@ -1,0 +1,3 @@
+import Button from './Button';
+
+export { Button };
