+++ src/rifui/components/atoms/forms/FormControl.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useState } from 'react';
+import React, { FC } from 'react';
 import { FormControl as MUIFormControl, FormControlProps as MUIFormControlProps } from '@material-ui/core';
 
 export interface IFormControlProps extends MUIFormControlProps { }
