+++ src/rifui/components/atoms/forms/index.ts
@@ -1,0 +1,13 @@
+import Checkbox from './Checkbox';
+import FormControl from './FormControl';
+import FormGroup from './FormGroup';
+import RangeSlider from './RangeSlider';
+import TextField from './TextField';
+
+export {
+  Checkbox,
+  FormControl,
+  FormGroup,
+  RangeSlider,
+  TextField
+};
