+++ src/rifui/components/atoms/index.ts
@@ -1,0 +1,11 @@
+import Logo from './Logo';
+import LogoFooter from './LogoFooter';
+import LogoNavbar from './LogoNavbar';
+
+export * from './buttons/index';
+export * from './forms/index';
+export {
+  Logo,
+  LogoFooter,
+  LogoNavbar,
+};
