+++ src/rifui/components/index.ts
@@ -1,0 +1,10 @@
+import Account from './Account';
+import LoginOption from './LoginOption';
+
+export * from './atoms/index';
+export * from './molecules/index';
+
+export {
+  Account,
+  LoginOption
+};
