+++ src/rifui/components/molecules/index.ts
@@ -1,0 +1,11 @@
+import Accordion from './Accordion';
+import FilterCheckboxCard from './FilterCheckboxCard';
+import LabeledCheckbox from './LabeledCheckbox';
+import RangeSliderWithInputs from './RangeSliderWithInputs';
+
+export {
+  Accordion,
+  FilterCheckboxCard,
+  LabeledCheckbox,
+  RangeSliderWithInputs
+};
