+++ src/rifui/index.ts
@@ -1,27 +1,5 @@
 // CSS files
 import './css/style.css';
 
-import Accordion from './components/Accordion';
-import Button from './components/atoms/buttons/Button';
-import Checkbox from './components/atoms/forms/Checkbox';
-import FormControl from './components/atoms/forms/FormControl';
-import FormGroup from './components/atoms/forms/FormGroup';
-import Logo from './Logo';
-import LogoFooter from './LogoFooter';
-import LogoNavbar from './LogoNavbar';
-import RangeSlider from './components/RangeSlider';
-import TextField from './components/atoms/forms/TextField';
-
 // Components
-export {
-    Accordion,
-    Button,
-    Checkbox,
-    FormControl,
-    FormGroup,
-    Logo,
-    LogoFooter,
-    LogoNavbar,
-    RangeSlider,
-    TextField
-};
+export * from './components/index';
